Acknowledge/transmission counters now size_t

+++ shared_mem.h
@@ -9,8 +9,8 @@
 struct SharedMem
 {
     std::array<addressRegistry::AddressStruct, addressRegistry::max_registry_size> addrRegistry;
-    int                                                                            acknowledgeCntr{0};
-    int                                                                            transmissionCntr{0};
+    size_t                                                                         acknowledgeCntr{0};
+    size_t                                                                         transmissionCntr{0};
     intptr_t                                                                       commandAddr;
     std::variant<int, double, std::array<double, 4>>                               commandVal;
     size_t                                                                         commandSize;
